Fix ion asset caching
ion asset caching was basically broken because sending the access token in
the query string would bust the cache. The initial fix was to allow the
token in the Authorization header, which fixed caching but triggered a
preflight request for every tile request (due to CORS), this adds way
too much overhead. To work around this CORS limitation, the ion servers
also support including the token in the `Accept` header, which is
non-standard but perfectly legal and does not trigger a pre-flight.

Also turned the private and static function `Resource._makeRequest` into
a prototype function so that I could override it `IonResource`.